fix: use number for lastModified and refactor storage tests
