+++ .vscode/settings.json
@@ -1,22 +1,21 @@
 {
   "workbench.colorCustomizations": {
-    "activityBar.activeBackground": "#8193f4",
-    "activityBar.background": "#8193f4",
+    "activityBar.activeBackground": "#d3a570",
+    "activityBar.background": "#d3a570",
     "activityBar.foreground": "#15202b",
     "activityBar.inactiveForeground": "#15202b99",
-    "activityBarBadge.background": "#fce2e6",
+    "activityBarBadge.background": "#e7f7f0",
     "activityBarBadge.foreground": "#15202b",
-    "commandCenter.border": "#e7e7e799",
-    "sash.hoverBorder": "#8193f4",
-    "statusBar.background": "#526bf0",
-    "statusBar.foreground": "#e7e7e7",
-    "statusBarItem.hoverBackground": "#8193f4",
-    "statusBarItem.remoteBackground": "#526bf0",
-    "statusBarItem.remoteForeground": "#e7e7e7",
-    "titleBar.activeBackground": "#526bf0",
-    "titleBar.activeForeground": "#e7e7e7",
-    "titleBar.inactiveBackground": "#526bf099",
-    "titleBar.inactiveForeground": "#e7e7e799"
-  },
-  "peacock.color": "#526bf0"
+    "commandCenter.border": "#15202b99",
+    "sash.hoverBorder": "#d3a570",
+    "statusBar.background": "#c78d49",
+    "statusBar.foreground": "#15202b",
+    "statusBarItem.hoverBackground": "#a97334",
+    "statusBarItem.remoteBackground": "#c78d49",
+    "statusBarItem.remoteForeground": "#15202b",
+    "titleBar.activeBackground": "#c78d49",
+    "titleBar.activeForeground": "#15202b",
+    "titleBar.inactiveBackground": "#c78d4999",
+    "titleBar.inactiveForeground": "#15202b99"
+  }
 }
